Previous triangle hit is now rechecked at the first ray of each frame.
In most cases, especially with no movement and large triangles, we will most likely hit the same triangle again. If we do we don't know whether there is some occluder now but we can atleast cull everything behind that before traversing the BVH. This improves performance by ~15% for first rays.