Reworked sky computation.
First and foremost this fixes a massive blunder where the extinction did not advance but simply sampled the same section all the time. Besides this I reworked a lot of things here and there. Ozone now uses a realistic distribution and sky intensity is now properly in sync with sun intensity. In total the computation is also a bit faster. The sky is also now using random offsets which adds a bit of noise in exchange for a better converged result at less iterations.