Reworked spatial resampling.
This is now more in line with ReSTIR. Sampling lights + sun is now much
more effective and triangle solid light computations are halved in the
spatial resampling pass. The white_noise calls need to be replaced
again for better performance. Unfortunately, the light sample size
doubled which reduces performance and costs memory space.