First step of the device refactor.
The goal is to store all GPU related code in the translation unit around
device.cu. Everything else is supposed to reside in C land which means
we introduce raytrace.c as the base file for RaytraceInstance. To make
this possible, we need a way to address constant memory from places
where they are not defined. To do this, the whole constant memory layout
is now defined by a struct and symbol memcpy now happens through wrapper
functions. These functions also help reduce the amount of code. This
whole procedure is broken into multiple commits as this commit is
already massive as all constant memory accesses needed to be changed.