Removed blue noise.
I tested this and the blue noise provided no benefit. This is a job for
future me to evaluate possible use of blue noise in the future. When I
introduced this blue noise it provided a huge improvement in image
quality. However, this is no longer the case. Since it has been almost 2
years since then, it is not really feasible to determine the cause.
Further, I plan to introduce a very fast random number generator in the
near future. Now that everything is white noise, the speed improvement
is going to be even more noticable I hope.

I must say this branch is a bit sad. So many things that were unchanged
in Luminary since the very early days have now been changed.