Added a proper shape to the top cloud layer.
As a result, the cloud system is almost done now. I left a note on
Cirrus clouds and that I currently don't see a solution for those. So
unless that changes in the next few days, I am now able to finish up my
work here so that I can finally mark this gigantic cloud rework as
complete.