Reworked binary BVH construction.
The numerical errors in the construction are a huge issue. The previous
version was a pretty lucky one that just barely scraped by and only
failed in some rare cases. The new version is slightly faster and has a
lot of error handling through which I encounter no failures anymore.