Added baked HDRI skybox computation.
The sky + clouds can now be baked into a texture. As a result, good
frametimes can be achieved even with clouds on. This approach has
several issues which were the reason I hadn't done this earlier. The
result is actually very good and the issues are less severe than
expected. This implementation is not quite complete yet as the sky needs
different wrapping modes for its different axis and cuda texture
creation memory leaks which is a severe issue since the skybox can be up
to a gigabyte in size (without mipmaps).