Updated zlib submodule.
I spent a lot of time investigating whether I could instead link against
zlib-ng. However, the final blocker is the MSVC requirement on Windows
which requires a MSVC build of zlib-ng due to the usage of unistd in
that library if GNU tools where used for building. It is a mess and not
quite satisfactory. I added some more notes and fixed some small CMake
things.