Added baseline optix raytrace setup.
It is quite insane how much setup is required for optix. However, I
think I can integrate this very nicely into my workflow in Luminary
since I will have access to my device table and the copying of the
table at every frame is cheap since it is device to device. The only
thing that I dislike right now is that I need to copy the ptx file to
the destination. A solution to avoid this would be to embed the ptx file
into the executable but that is something that I will do towards the end
of this project.