Implemented working optix raytracing.
This implementation is working but there is a lot left to do. There is
no alpha cutoff currently so that needs to be worked on. Also the
construction of the Luminary BVH reorders the triangles which is an
issue due to triangle ID mismatches. To fix this I need to rearrange the
index buffer aswell. The performance right now is around 2x of what
the Luminary BVH provides. I think this is a huge W for the Luminary BVH
but lets see what the final implemenation's performance will look like.