Added transparency support and fixed triangle reordering for OptiX.
Unfortunately, I dont think there is a way of gathering traversal
information for OptiX that I could use for the heat shading mode.
Sorting traversal elements is also still broken. Finally, both BVH and
OptixRT need a refactoring so that they are both similar in API and are
only computed when actually used. I already like the OptixRT API but I
want to move the whole pipeline building into one function. I wont need
access to the modules or groups afterwards anyway so they dont need to
be stored in the RaytraceInstance.