Added support for gamma encoding in textures.
This was the cause of my issues with normal maps. Currently, only the
simple gAMA chunk is parsed. GIMP for some reason decides to not follow
the PNG standard and hence the gAMA chunk is broken in GIMP exports.
However, it seems that GIMP by default just exports with linear encoding
because interpreting the image as linear works without any issues. Maybe
trying to parse the ICC Profile that GIMP includes could be a way of
alleviating this issue in the future.