Fixed a bug in light resampling with very few lights.
Triangle lights would be brighter if the sun is present if only few
triangle lights are present. Turns out that if I allow resampling to
resample against the sample sample then I obtain the correct result.
Intuitively, this makes no sense since this should introduce a bias if
for example the same light is always selected. However, I guess in that
case the ratio of sum of weights and number of samples stays the same so
it is fine.