Added preliminary DMM implementation.
It works now without crashing but the result still does not look right
despite the fact that I currently only try to zero the displacement.
Also I will need to add some additional features that with for example
deciding whether OMM and DMM are allowed to be used at the same time.