Removed terminator handling and made post hit position shift more robust.
The only valid direction to shift the position is along the ray. Any
other direction can push the position through other geometry which
causes light leaks. This change fixes many cases of lights leaks.
However, the shadow terminator is now also a bit more of a problem
again. The shadow terminator problem is after all not solvable and I
think the way it is now is the sanest approach. If something suffers
from huge shadow terminator issues, then the mesh should be refined.
After all, the shadow terminator issue is simply an issue where what we
desire the mesh to look like does not match what we obtain, yet what we
obtain is now more accurate.