Further work on ReSTIR.
This now basically works and the reduction in noise is massive.
However, the bias is also very substantial. The bias is introduced due
to essentially setting the target pdf to zero for lights that were
occluded. This mismatch then causes a bias when resampling the
neighbours. This is addressed in the papers but implementing the bias
correction is yet to be done.