Fixed the spatiotemporal resampling in ReSTIR.
The issue was not directly bias related (or I guess you can call it that
if you want) but rather I messed up entirely by preselecting the pixel's
current sample as the start sample. Without doing this if I were to only
do spatiotemporal resampling in one go I fear I would sample the new
samples too rarely. Hence I decided to swap to separate temporal and
spatial resampling as in the original paper. This now works without
introducing any major bias (though this may be biased but I have not
checked the math, this is a bit of a delicate topic for this algorithm
anyway). From here on it becomes important to improve the performance of
the resampling both in terms of computational cost and by reusing
previous confirmed to be visible samples more often.