Removed visibility reuse in ReSTIR.
So I now understand how to alleviate BIAS issues in ReSTIR. However,
this involves evaluating the target pdf of the selected sample for all
origin pixels. This means we need to reevaluate visibility to remove
BIAS. This is not acceptable and hence I need to drop it. I will attempt
including a full BRDF evaluation though which could improve the sample
quality some more, for example, by excluding lights that are on the other
side of the hit normal.