Fixed some more bugs in ReSTIR and documented implementation.
Both temporal and spatial resampling will be removed as they seem to not
be worth it. Ultimately, my ReSTIR implementation is going to just be
simple WRS, even simpler than the implementation I started with.
However, this implementation performs pretty well. From here on
performance will be the main concern and possible preparations towards
GRIS.