Fog now also uses ReSTIR.
I had initially planned to scrap the fog entirely, however, I actually
got somewhat okay results with this now. There is clearly something
wrong with the extinction but once that is fixed I think the fog could
be quite decent (I am really back and forth with this one over the last
years). Also I had the idea to maybe rename the light eval data to a per
iteration G-Buffer which could be used by the actual shading kernels,
not sure yet whether that is useful.