Reworked ocean surface intersection.
The new implementation allows for fine detail and multiple scattering
on the ocean surface when close to the camera. This new approach is also
a lot more expensive though. Some features like the animated ocean where
removed since that roughly doubles performance and that feature was
never used by me anyway. The old intersection test is still used, albeit
in an overhauled way, for far away ocean intersection tests. There are
both performance and qualitative improvements to be done for the ocean
surface but as it stands this is fine for now so I can focues on
finishing the volumetrics.