Added state handler system and fixed task based crash.
The crash was very tedious to figure out. Ultimately, the fix was to
allocate one more pixel to each thread. This bug only now appeared
because dense fog can actually maximize the number of tasks very
consistently. The balance kernel also had some bugs but they were not
the source of the problem.