Added phase function importance sampling and fixed sphere sampling.
The phase function importance sampling is yet to be used in more than
just fog or ocean. The sphere sampling was bugged in the BRDF
implementation. This new implementation is definitely correct now but I
am unsure whether any old implementations are still wrong, I will have
to look into this.