Replaced all uses of RGBAhalf with RGBF/RGBAF.
I made the switch to half precision color values over a year ago. Back
then memory consumption was a huge concern as Luminary at that point had
close to no volumetrics and the cost of rendering was rather small.
Nowadays, high resolution are not feasible due to the high rendering
cost, hence per pixel memory consumption is now less important. Further,
the half precision colors failed quite often because the sun is just so
bright. This change will put more of an emphasis on image quality.