Fixed triangle solid angle scale.
This one is truly one of nature mysteries. In my head I remember reading
a factor of 4, yet I can't find a single reference that actually claims
this. All reference say 2 which probably correct then. Funny that
essentially for the last 2 years everything direct lighting was wrong
because of this.