Refactored shading task structs.
They could now all be unified to a single shading task struct but I want
to keep it this way due to tradition and because it makes it easier to
add data to the struct of a specific shading task if needed, i.e., if
the shading task requirements diverge again. It is honestly quite
interesting to see this given that this was one of the very early things
I introduced when I reworked the kernel flow in 2021.