Improved bench output and fixed some bugs in the process.
The new bench output makes it possible to see what is currently being
done. I tried a progress bar but I noticed quickly that I cannot
implement that for most functions so it seemed pointless. Also there
were some issues that popped up because I changed the output of the the
offline renderer.