Refactored the ReSTIR sample evaluation to be coherent for surface and volumes.
This actually improves performance a lot in scenes with volumes.
Previously it would also have to execute this twice essentially, once
for all threads in the warp with a surface, then for all threads in the
warp with a volume, now it does it for all threads at the same time.