Moved the solid angle of the light into the sampling pdf.
This is actually quite interesting as there are two approaches and they
are both equal. My old approach was, pretend we would perfectly
integrate over the light which would result in a contribution factor of
solid angle over total domain area. The new approach is we consider
taking a random direction towards the light (as we actually do) and
account for the probability of sampling this ray which is given by the
solid angle. They are both producing the exact same result. The new
approach is more true to the actual nature of path tracing and at least
in theory means we won't need to compute the solid angle in the shading
step. However, we get the solid angle for free anyway so there is pretty
much no saving through this. This is really nothing more but a matter of
philosophy I guess. Also I saw that I don't export the solid angle when
we are inside a sphere so I fixed that.