Light pass transparency roulette is now almost unbiased.
There was this issue where I only multiplied the pdf of the roulette
into the current record which meant that light rays in later bounces
would not be able to account for this. Ultimately,
this is probably not 100% unbiasable because I think you would always
need hindsight to correctly weigh things. For example, say at bounce 0 we
shoot a light ray that hits transparency for two passes in a row. First
bounce we continue, now comes the roll at the second pass. The question
is: What is the weight after this? Is it 4 because 2 * 2 or is it only 2
because a potential ray in the first bounce could have also hit a
transparency after 1 pass meaning that the first decision in the
roulette is no longer contributing to the actual pdf. However, if that
light ray in the first bounce (which was never actually computed)
wouldn't hit a transparent surface, then the probability would actually
be 3/4 because the only way of still being blocked would be to roll
twice on continue in the light ray in the 0th bounce. With all that said
though, even in scenes where light rays have to pass through multiple
layers of transparencies, this approach now converges to a visually
perfect result. Hence I think I can accept this.