Merged GBuffer, ReSTIR and shading kernels.
This completely removes the need for the GBuffer buffer. All in all we
save lots of memory traffic that way. In my test scene frame times went
from 345ms to 320ms which is quite nice. The amount of changes seems
massive but it is just a lot of copying around especially because I had
some cyclic include dependencies.