Merged volume event kernels.
This reduced frametimes in my test scene from 309ms to 275ms. This is
mostly because that scene was heavily limited by the ocean intersection
computation. We now compute the volume path one less time so that is why
we gain so much. Additionally, I save loading the TraceTasks and results
once.