Fixed HDRI generation always generating the same frame.
I forgot that the new random numbers would not work with the HDRI
generation since I don't set the temporal frames in there. I simply went
for the maintainable approach to fix this which means that each HDRI
frame is now generated in sequence. This also has the advantage that I
can now display progress state.