Fixed fireflies caused by russian roulette.
This is the 1000th commit. In standard Luminary fashion, fireflies are
the bain of our existence and the only solution is fucking some things
up. But hey it seems fine now. However, the white furnace now seems to
be wrong again. But how I wonder since there is literally no reason for
that. Anyway, all the actual scenes look amazing.