All kernels now use the same launch bounds.
The nvidia compiler is going to be better overall at determining optimal
register count that I will. This will also fix issues with other
architectures that I can't test. We keep the threads per block
launchbound as that can still be helpful for the compiler. Overall I am
seeing a ~5% performance gain here which is quite decent.