Removed light ray condition to remove some fireflies.
Caustics like that are not properly sampleable anyway, this check never
made any sense. If we want this we should add back the BOUNCE_LIGHTING
flag but I don't want to do that because those kind of paths would never
be sampled properly anyway and would just cause fireflies. I am fine
with this until I do proper GI importance sampling which might be never
:)