Changed the caustic sampling domain.
This now manages to sample the caustics reasonably well. The only issue
I see right now is how amplitude==0 provides a very different result, it
is very close in brightness to the non caustic lighting. If the
amplitude!=0 then the ocean is much darker. I think my weights are
fucked (not surprised tbh).