Geometry lights are now only sampled through DL.
It seems to work better than what I had before, however, it requires a
clamp to the roughness. From here I think I want to clean this up and do
the same for non geometry lights. Also I need a special case for
roughness 0 because that might aswell be quite nice. On top of that I
think the best way to solve my ocean problem is by merging the ocean
into the geometry optix kernel and let it do DL aswell.