Fixed an issue with transparent pass through rays.
Refraction DL still has issues and it is not capable of sampling pass
through rays. However, even if it was capable, it shouldn't do it as we
would otherwise run into double counting. To fix this I added this flag
as the real issue simply comes from the fact that we don't do DL at the
camera path vertex.