Implemented construction of light tree.
From here I need to convert it into something that can be traversed
quickly on the GPU. This was just thrown together using my BVH
implementation. It manages to not crash, that is all that I know for
now. I am sure I am going to have lots of fun debugging whatever lies
ahead of me.