Fixed geometry -> light ptrs.
I was still assigning them before the light tree construction which is bogus
given that the light id change during the construction. I did this to
fix an issue where the light intensity displayed in the lights mode was
completely messed up which was ultimately only a display issue that
didnt affect sampling. However, the broken links caused a bug with the
shadowing that I originally thought to be an Optix bug but now I just
have no idea how these broken links caused this. Anyway I am very happy
that this is fixed now because this means that light tree are pretty
much in a ready state now. I am not completely happy with the noise yet,
especially because to me it seems that the BSDF rays are sometimes very
noisy and I am not sure why.