Bridges now uses HG phase function with g=0.85.
The small amount of backscattering in my normal phase functions causes
lots of noise and fireflies. I have two choices: Have a lot of noise
that ruins the renders or add a little bit of bias by having a different
phase function. I took the second option. This is clearly an area where
improvements would be nice in the future. As expected, the topic of
volumes is no where near solved.