Renderer now computes 1 tile per iteration due to CUDA queue limits.
CUDA has a limit where you can't have more than roughly 1000 kernels
queued over all streams. You will find plenty of people that say that
this is per stream, however, I found that this is an overall limit and
does not just apply per stream. This here is the best solution, though
it is clear that at high depths you will still run into this issue, but
that is nothing new.