Changed the behaviour of the device mask at renderer creation.
I spend some thought on supporting disabling devices and
unloaded/reloading the device memory. I came to the conclusion that this
would obviously be possible but it would require a lot of work and the
system would be somewhat complicated and possibly error prone. So
ultimately, by simply using the device mask to only selectively create
the devices. The issue that I am dealing with in the first place is
mostly addressed and I can keep working with a very simple memory
management system where a available device being disabled is not major
use case concern anymore.